Store `rabbitmq_aws` credentials in an ETS table
This replaces the `rabbitmq_aws` gen_server which wrapped `httpc` with
an ETS table. The gen_server updated credentials when they expired, but
it also blocked for the duration of any request, so it was not suitable
for making multiple in-flight requests. Instead of using a gen_server as
the consistency control we can use a lock lock so that only one process
attempts to refresh credentials at a time. The ETS table can then be
read by any processes independently and requests can be sent directly
rather than going through the gen_server.

This change also switches the HTTP client from `httpc` to `gun`.

Co-authored-by: Michael Davis <mcrdavis@amazon.com>